added shared and wait methods, update index.ts
