+++ src/web/methods/index.ts
@@ -1,7 +1,11 @@
 import * as interactions from './interactions.methods';
 import * as navigation from './navigation.methods';
+import * as shared from './shared.methods';
+import * as wait from './wait.methods';
 
 export const methods = {
   interactions,
   navigation,
+  shared,
+  wait,
 };
